+++ tests/conftest.py
@@ -172,6 +172,12 @@ def fixture_base_config_dict() -> dict:
             "database": "sqlite:///",
             "root": "path/to/root",
         },
+        "email_base_settings": {
+            "sll_port": 465,
+            "smtp_server": "smtp.gmail.com",
+            "sender_email": "test@gmail.com",
+            "sender_password": "",
+        },
     }
 
 
@@ -1126,6 +1132,12 @@ def fixture_context_config(
         "bed_path": str(cg_dir),
         "delivery_path": str(cg_dir),
         "hermes": {"deploy_config": "hermes-deploy-stage.yaml", "binary_path": "hermes"},
+        "email_base_settings": {
+            "sll_port": 465,
+            "smtp_server": "smtp.gmail.com",
+            "sender_email": "test@gmail.com",
+            "sender_password": "",
+        },
         "demultiplex": {
             "run_dir": "tests/fixtures/apps/demultiplexing/flowcell_runs",
             "out_dir": "tests/fixtures/apps/demultiplexing/demultiplexed-runs",
@@ -1162,6 +1174,12 @@ def fixture_context_config(
             "service_account": "SERVICE",
             "service_account_auth_file": "trailblazer-auth.json",
             "host": "https://trailblazer.scilifelab.se/",
+        },
+        "gisaid": {
+            "binary_path": "/path/to/gisaid_uploader.py",
+            "log_dir": "/path/to/log",
+            "submitter": "s.submitter",
+            "logwatch_email": "some@email.com",
         },
         "lims": {
             "host": "https://lims.scilifelab.se",